requirements.txt
bcrypt==3.1.7
PyJWT==2.6.0
requests==2.22.0
